Move domain classes to domain package
Easy organisation.
