+++ .idea/modules.xml
@@ -2,6 +2,7 @@
 <project version="4">
   <component name="ProjectModuleManager">
     <modules>
+      <module fileurl="file://$PROJECT_DIR$/.idea/modules/WaystoneWarps.WorldWideWarps.main.iml" filepath="$PROJECT_DIR$/.idea/modules/WaystoneWarps.WorldWideWarps.main.iml" />
       <module fileurl="file://$PROJECT_DIR$/.idea/modules/WorldWideWarps.main.iml" filepath="$PROJECT_DIR$/.idea/modules/WorldWideWarps.main.iml" />
       <module fileurl="file://$PROJECT_DIR$/.idea/modules/WorldWideWarps.test.iml" filepath="$PROJECT_DIR$/.idea/modules/WorldWideWarps.test.iml" />
       <module fileurl="file://$PROJECT_DIR$/.idea/modules/dev.mizarc.WorldWideWarps.main.iml" filepath="$PROJECT_DIR$/.idea/modules/dev.mizarc.WorldWideWarps.main.iml" />
