+++ src/main/kotlin/dev/mizarc/worldwidewarps/Teleporter.kt
@@ -1,5 +1,7 @@
 package dev.mizarc.worldwidewarps
 
+import dev.mizarc.worldwidewarps.domain.PlayerRepository
+import dev.mizarc.worldwidewarps.domain.Warp
 import net.kyori.adventure.text.Component
 import net.kyori.adventure.text.format.TextColor
 import org.bukkit.Bukkit
