+++ src/main/kotlin/dev/mizarc/worldwidewarps/listeners/BedDestructionListener.kt
@@ -4,7 +4,7 @@ import org.bukkit.block.data.type.Bed
 import org.bukkit.event.EventHandler
 import org.bukkit.event.Listener
 import org.bukkit.event.block.BlockBreakEvent
-import dev.mizarc.worldwidewarps.HomeRepository
+import dev.mizarc.worldwidewarps.domain.HomeRepository
 import dev.mizarc.worldwidewarps.Position
 import org.bukkit.block.Block
 import org.bukkit.event.block.BlockExplodeEvent
