+++ src/main/kotlin/dev/mizarc/worldwidewarps/listeners/TeleportCancelListener.kt
@@ -5,7 +5,7 @@ import net.kyori.adventure.text.format.TextColor
 import org.bukkit.event.EventHandler
 import org.bukkit.event.Listener
 import org.bukkit.event.player.PlayerMoveEvent
-import dev.mizarc.worldwidewarps.PlayerRepository
+import dev.mizarc.worldwidewarps.domain.PlayerRepository
 
 class TeleportCancelListener(val playerRepository: PlayerRepository): Listener {
